refactor: move utilities to owning modules
