+++ src/cli-schema/cli-config.ts
@@ -2,7 +2,7 @@ import type { CliFlags } from '@agent-device/contracts/command';
 import fs from 'node:fs';
 import path from 'node:path';
 import { AppError } from '@agent-device/kernel/errors';
-import { mergeDefinedFlags } from '../utils/merge-flags.ts';
+import { mergeDefinedFlags } from './merge-flags.ts';
 import { type FlagKey } from '../commands/cli-grammar/flag-types.ts';
 import { expandUserHomePath, resolveUserPath } from '@agent-device/host-kit/file';
 import {
