+++ src/cli-schema/merge-flags.ts
