+++ src/commands/interaction/runtime/truncate-utf8.ts
