+++ src/commands/observability/string-enum.ts
