+++ src/core/__tests__/device.test.ts
