+++ src/core/dispatch-payload.ts
@@ -1,6 +1,6 @@
 import { promises as fs } from 'node:fs';
 import { AppError } from '@agent-device/kernel/errors';
-import { resolvePayloadInput } from '../utils/payload-input.ts';
+import { resolvePayloadInput } from './payload-input.ts';
 import type { JsonObject } from '@agent-device/contracts/client';
 
 export async function readNotificationPayload(payloadArg: string): Promise<JsonObject> {
