+++ src/core/parse-timeout.ts
