+++ src/daemon/handlers/session-app-deployment.ts
@@ -11,7 +11,7 @@ import { expireRefFrame } from '../ref-frame.ts';
 import type { BindDeviceRuntime, InspectDeviceRuntimeFacts } from '../request-runtime-binding.ts';
 import { SessionStore } from '../session-store.ts';
 import type { DaemonRequest, DaemonResponse, SessionState } from '../types.ts';
-import { resolvePayloadInput } from '../../utils/payload-input.ts';
+import { resolvePayloadInput } from '../../core/payload-input.ts';
 import { resolveDeployResultTarget } from '../result-serialization.ts';
 import { withSuccessText } from '@agent-device/kernel/success-text';
 import { recordSessionAction } from '../session-action-recorder.ts';
