+++ src/daemon/session-event-action-presentation.ts
@@ -3,7 +3,7 @@ import { DEVICE_ROTATIONS } from '@agent-device/contracts/device';
 import { BACK_MODES } from '@agent-device/contracts/back-mode';
 import { TV_REMOTE_BUTTONS } from '@agent-device/contracts/tv-remote';
 import { PUBLIC_COMMANDS } from '../command-catalog.ts';
-import { isKeyboardAction } from '../utils/keyboard-actions.ts';
+import { isKeyboardAction } from '../core/keyboard-actions.ts';
 import {
   compactSessionEventDetails as compactDetails,
   readSessionEventBoolean as readBoolean,
